+++ lang/en/attendance.php
@@ -33,12 +33,14 @@ $string['Pfull'] = 'Present';
 $string['acronym'] = 'Acronym';
 $string['add'] = 'Add';
 $string['addmultiplesessions'] = 'Multiple sessions';
+$string['addwarning'] = 'Add warning';
 $string['addsession'] = 'Add session';
 $string['adduser'] = 'Add user';
 $string['all'] = 'All';
 $string['allcourses'] = 'All courses';
 $string['allpast'] = 'All past';
 $string['allsessions'] = 'All sessions';
+$string['atriskreport'] = 'At-risk report';
 $string['attendance:addinstance'] = 'Add a new attendance activity';
 $string['attendance:canbelisted'] = 'Appears in the roster';
 $string['attendance:changeattendances'] = 'Changing Attendances';
@@ -104,6 +106,9 @@ The sessions begin on the date of the base session and continue until the \'repe
 $string['createonesession'] = 'Create one session for the course';
 $string['days'] = 'Days';
 $string['defaultdisplaymode'] = 'Default display mode';
+$string['defaultwarnings'] = 'Default warning set';
+$string['defaultwarningsettings'] = 'Default warning settings';
+$string['defaultwarningsettings_help'] = 'These settings define the defaults for all new warnings';
 $string['defaults'] = 'Defaults';
 $string['defaultsessionsettings'] = 'Default session settings';
 $string['defaultsessionsettings_help'] = 'These settings define the defaults for all new sessions';
@@ -115,6 +120,7 @@ $string['defaultsubnet_help'] = 'Attendance recording may be restricted to parti
 $string['defaultview'] = 'Default view on login';
 $string['defaultview_desc'] = 'This is the default view shown to teachers on first login.';
 $string['delete'] = 'Delete';
+$string['deletewarningconfirm'] = 'Are you sure you want to delete this warning?';
 $string['deletedgroup'] = 'The group associated with this session has been deleted';
 $string['deletehiddensessions'] = 'Delete all hidden sessions';
 $string['deletelogs'] = 'Delete attendance data';
@@ -134,8 +140,35 @@ $string['downloadtext'] = 'Download in text format';
 $string['duration'] = 'Duration';
 $string['editsession'] = 'Edit Session';
 $string['edituser'] = 'Edit user';
+$string['emailcontent_default'] = 'Hi %userfirstname%,
+Your attendance in %coursename% %attendancename% has dropped below %warningpercent% and is currently %percent% - we hope you are ok!
+
+To get the most out of this course you should improve your attendance, please get in touch if you require any further support.';
+$string['emailcontent'] = 'Email content';
+$string['emailcontent_help'] = 'When a warning is sent to a student, it takes the email content from this field. The following wildcards can be used:
+<ul>
+<li>%coursename%</li>
+<li>%userfirstname%</li>
+<li>%userlastname%</li>
+<li>%userid%</li>
+<li>%warningpercent%</li>
+<li>%attendancename%</li>
+<li>%cmid%</li>
+<li>%numtakensessions%</li>
+<li>%points%</li>
+<li>%maxpoints%</li>
+<li>%precent%</li>
+</ul>';
+
+$string['emailsubject'] = 'Email subject';
+$string['emailsubject_help'] = 'When a warning is sent to a student, it takes the email subject from this field.';
+$string['emailsubject_default'] = 'Attendance warning';
+$string['emailuser'] = 'Email user';
+$string['emailuser_help'] = 'If checked, a warning will be sent to the student.';
 $string['emptyacronym'] = 'Empty acronyms are not allowed. Status record not updated.';
 $string['emptydescription'] = 'Empty descriptions are not allowed. Status record not updated.';
+$string['enablewarnings'] = 'Enable warnings';
+$string['enablewarnings_desc'] = 'This allows a warning set to be defined for an attendance and email notifications to users when attendance drops below the configured threshold. <br/><strong>WARNING: This is a new feature and has not been tested extensively. Please use at your own-risk and provide feeback in the moodle forums if you find it works well.</strong>';
 $string['endofperiod'] = 'End of period';
 $string['endtime'] = 'Session end time';
 $string['enrolmentend'] = 'User enrolment ends {$a}';
@@ -175,6 +208,7 @@ $string['includeremarks'] = 'Include remarks';
 $string['incorrectpassword'] = 'You have entered an incorrect password and your attendance has not been recorded, please enter the correct password.';
 $string['indetail'] = 'In detail...';
 $string['invalidaction'] = 'You must select an action';
+$string['invalidemails'] = 'You must specify addresses of existing user accounts, could not find: {$a}';
 $string['invalidsessionenddate'] = 'This date can not be earlier than the session date';
 $string['invalidsessionendtime'] = 'The end time must be greater than start time';
 $string['invalidstatus'] = 'You have selected an invalid status, please try again';
@@ -217,9 +251,15 @@ $string['noofdayspresent'] = 'No of days present';
 $string['nosessiondayselected'] = 'No Session day selected';
 $string['nosessionexists'] = 'No Session exists for this course';
 $string['nosessionsselected'] = 'No sessions selected';
+$string['warningdeleted'] = 'Warning deleted';
+$string['warningdesc'] = 'These warnings will be automatically added to any new attendance activities. If more than one warning is triggered at exactly the same time, only the warning with the lower warning threshold will be sent.';
+$string['warnings'] = 'Warnings set';
+$string['warningupdated'] = 'Updated warnings';
+$string['notifytask'] = 'Send warnings to users';
 $string['notfound'] = 'Attendance activity not found in this course!';
 $string['notmember'] = 'not&nbsp;member';
 $string['noupgradefromthisversion'] = 'The Attendance module cannot upgrade from the version of attforblock you have installed. - please delete attforblock or upgrade it to the latest version before isntalling the new attendance module';
+$string['numsessions'] = 'Number of sessions';
 $string['olddate'] = 'Old date';
 $string['onlyselectedusers'] = 'Export specific users';
 $string['overallsessions'] = 'Over all sessions';
@@ -384,10 +424,16 @@ $string['tempusermerge'] = 'Merge temporary user';
 $string['tempusers'] = 'Temporary users';
 $string['tempusersedit'] = 'Edit temporary user';
 $string['tempuserslist'] = 'Temporary users';
+$string['thirdpartyemailsubject'] = 'Attendance warning';
+$string['thirdpartyemailtext'] = '{$a->firstname} {$a->lastname} attendance within {$a->coursename} {$a->aname} is lower than {$a->warningpercent} ({$a->percent})';
+$string['thirdpartyemailtextfooter'] = 'You are receiving this because the teacher of this course has added your email to the recipient’s list';
+$string['thirdpartyemails'] = 'Notify other users';
+$string['thirdpartyemails_help'] = 'List of other users who will be notified. (requires the capability mod/attendance:viewreports)';
 $string['thiscourse'] = 'This course';
 $string['time'] = 'Time';
 $string['timeahead'] = 'Multiple sessions that exceed one year cannot be created, please adjust the start and end dates.';
 $string['to'] = 'to:';
+$string['triggered'] = 'First triggered';
 $string['tuseremail'] = 'Email';
 $string['tusername'] = 'Full name';
 $string['unknowngroup'] = 'Unknown group';
@@ -400,6 +446,12 @@ $string['variable'] = 'variable';
 $string['variablesupdated'] = 'Variables successfully updated';
 $string['versionforprinting'] = 'version for printing';
 $string['viewmode'] = 'View mode';
+$string['warnafter'] = 'Number of sessions taken before warning';
+$string['warnafter_help'] = 'Warnings will only be triggered when the user has had their attendance taken for at least this number of sessions.';
+$string['warningfailed'] = 'You cannot create a warning that uses the same percentage and number of sessions.';
+$string['warningpercent'] = 'Warn if percentage falls under';
+$string['warningpercent_help'] = 'A warning will be triggered when the overall percentage falls below this number.';
+$string['warningthreshold'] = 'Warning threshold';
 $string['week'] = 'week(s)';
 $string['weeks'] = 'Weeks';
 $string['youcantdo'] = 'You can\'t do anything';